Moved discord roles import logic inside a hook
